% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_missing.R
\name{append_missing}
\alias{append_missing}
\title{Append missing objects to the irm arrays and eigen decomposition objects.
Update irm_key_lookup matrix.}
\usage{
append_missing(epimodel)
}
\value{
epimodel with complete irm and eigen decomposition arrays.
}
\description{
Append missing objects to the irm arrays and eigen decomposition objects.
Update irm_key_lookup matrix.
}
