\name{BDAepimodel-package}
\alias{BDAepimodel-package}
\alias{BDAepimodel}
\docType{package}
\title{
\packageTitle{BDAepimodel}
}
\description{
\packageDescription{BDAepimodel}
}
\details{

\packageDESCRIPTION{BDAepimodel}
\packageIndices{BDAepimodel}
}
\author{
\packageAuthor{BDAepimodel}

\packageMaintainer{BDAepimodel}
}
\references{
}

\keyword{ package }
\seealso{
 \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
}
