% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_fb_mats.R
\name{build_fb_mats}
\alias{build_fb_mats}
\title{Construct the forward-backward matrices in the forward pass of the FB
algorithm.}
\usage{
build_fb_mats(epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
list containing the FB matrices for the stochastic FB algorithm.
}
\description{
Construct the forward-backward matrices in the forward pass of the FB
algorithm.
}

