% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_irm.R
\name{build_irm}
\alias{build_irm}
\title{Construct an environment of rate matrices and another environment for the
corresponding eigen decompositions.}
\usage{
build_irm(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}
}
\value{
environment with rate matrices
}
\description{
Each rate matrix in the array has dimension \code{num_states x num_states}.
}

