% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_new_irms.R
\name{build_new_irms}
\alias{build_new_irms}
\title{Build a new array of rate matrices}
\usage{
build_new_irms(epimodel, params)
}
\arguments{
\item{epimodel}{}

\item{params}{names vector of parameters}
}
\value{
epimodel object with updated rate array
}
\description{
Build a new array of rate matrices
}

