% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tpm_seqs.R
\name{build_tpm_seqs}
\alias{build_tpm_seqs}
\title{Construct the sequences of transition probability matrix products.}
\usage{
build_tpm_seqs(epimodel)
}
\value{
modify the .tpms and .tpm_products lists in the .epimodel environment
  and reset the .tpms_to_build vector
}
\description{
Updates the .tpms and .tpm_products lists. Resets the .tpms_to_compute vector
to be all FALSE. The elements of this vector needing to be updated are then 
set to true as each subject level trajectory is re-drawn, and when parameters
are updated.
}
