% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildEigenArray_SIR}
\alias{buildEigenArray_SIR}
\title{Update eigen values, vectors, and inverse matrices analytically for the SIR 
model}
\usage{
buildEigenArray_SIR(real_eigenvals, imag_eigenvals, eigenvecs, inversevecs,
  irm_array, n_real_eigs, initial_calc)
}
\arguments{
\item{eigenvecs}{array of eigenvectors with third dimension corresponding to
irms}

\item{inversevecs}{array of inverses of matrices of eigenvectors}

\item{irm_array}{array of rate matrices}

\item{eigenvals}{matrix of eigenvalues with columns corresponding to irms}

\item{complex_eigs}{logical vector indicating whether any eigen values are complex}

\item{first_calc}{boolean indicating whether this is the first time that the 
eigen decomposition is being computed}
}
\value{
Updated eigenvalues, eigenvectors, and inverse matrices
}
\description{
Update eigen values, vectors, and inverse matrices analytically for the SIR 
model
}
