% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildFBMats}
\alias{buildFBMats}
\title{Construct forward-backward matrices}
\usage{
buildFBMats(fb_mats, tpm_prods, emit_mat, initdist, obs_time_inds)
}
\arguments{
\item{fb_mats}{array of FB matrices to be updated}

\item{tpm_prods}{array of tpm products}

\item{emit_mat}{matrix of emission probabilities}

\item{initdist}{vector of initial state probabilities}

\item{obs_time_inds}{vector of observation time indices}
}
\value{
Updated array of FB matrices
}
\description{
Construct forward-backward matrices
}

