% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_likelihoods.R
\name{calc_likelihoods}
\alias{calc_likelihoods}
\title{Calculate likelihood for the population level trajectory and the measurement
process.}
\usage{
calc_likelihoods(epimodel, pop = TRUE, obs = TRUE, epimodel_envir = FALSE,
  log = FALSE)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{pop}{TRUE/FALSE for whether to compute the population-level
log-likelihood}

\item{obs}{TRUE/FALSE for whether to compute the likelihood of the
measurement process.}

\item{log}{TRUE/FALSE for whether to return the log-likelihood or likelihood.}
}
\value{
vector with log-likelihood of the population level trajectory and
  measurement process log-likelihood.
}
\description{
Calculate likelihood for the population level trajectory and the measurement
process.
}

