% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pop_likelihood.R
\name{calc_pop_likelihood}
\alias{calc_pop_likelihood}
\title{Calculate likelihood for the population level trajectory}
\usage{
calc_pop_likelihood(epimodel, log = TRUE)
}
\arguments{
\item{log}{TRUE/FALSE for whether to return the log-likelihood or likelihood.}

\item{params}{optional vector of parameters if values other than the current
values should be used (e.g. in a transition kernel)}
}
\value{
vector with log-likelihood of the population level trajectory and
  measurement process log-likelihood.
}
\description{
Calculate likelihood for the population level trajectory
}
