% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_subj_likelihood.R
\name{calc_subj_likelihood}
\alias{calc_subj_likelihood}
\title{Calculate the likelihood for a subject-level trajectory based on a
time-inhomogeneous CTMC.}
\usage{
calc_subj_likelihood(epimodel, subject, subj_ID, log = TRUE)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{subj_ID}{string for subject ID}

\item{log}{TRUE/FALSE for whether to return the log-likelihood or likelihood.}
}
\value{
likelihood or log-likelihood
}
\description{
Calculate the likelihood for a subject-level trajectory based on a
time-inhomogeneous CTMC.
}

