% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detect_structure.R
\name{detect_structure}
\alias{detect_structure}
\title{Detects whether there is structure in the stochastic epidemic model that can
be leveraged.}
\usage{
detect_structure(epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
objects in the epimodel environment: absorbing states (or FALSE if
  none), and with TRUE/FALSE for whether the model is progressive.
}
\description{
Detects whether the model has absorbing states (exit from such states is
impossible), and whether the model is progressive. If there is an absorbing
state, no additional path sampling is needed once a subject has entered that
state. If the model is progressive, it only moves forward through states and
never returns to a previous state. Therefore, if the model is progressive and
has an absorbing state, only the transition times in intervals where the
state has changed need to be resampled.
}

