% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_trajec.R
\name{draw_trajec}
\alias{draw_trajec}
\title{Sample a subject-level trajectory.}
\usage{
draw_trajec(epimodel, subject, iter)
}
\arguments{
\item{iter}{iteration number}
}
\value{
updated configuration matrix and observation matrix objects in the
  epimodel environment.
}
\description{
Proceeds in four steps: 1) Sample disease status at observation times using
the stochastic forward-backward algorithm. 2) Sample the infection status at 
transition times of other subjects. 3) Sample exact times of transition 
within inter-event intervals where needed. 4) Accept/Reject the proposed 
trajectory via Metropolis-Hastings.
}

