% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval_rates.R
\name{eval_rates}
\alias{eval_rates}
\title{Evaluate the rate functions at the current state and the parameter values}
\usage{
eval_rates(rate_fcns, state, params)
}
\arguments{
\item{rate_fcns}{list of rate functions created by extract_rate_fcns}

\item{state}{named vector or matrix of states, with names corresponding to
compartment names}
}
\value{
vector of rates
}
\description{
Evaluate the rate functions at the current state and the parameter values
}

