% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_epimodel_auxilliary.R
\name{extract_rate_fcns}
\alias{extract_rate_fcns}
\title{Extracts a list of functions for computing flow rates.}
\usage{
extract_rate_fcns(rates, compartments, params)
}
\arguments{
\item{rates}{character vector of rates of flow rates between compartments
with compartments corresponding to names of elements of \code{init_state}.}

\item{compartments}{vector of compartment names}

\item{params}{vector of parameter names}
}
\value{
Returns a list of functions to compute rates.
}
\description{
Extracts a list of functions for computing flow rates.
}
\examples{
rates <- c("beta * S * I", "mu * I")
compartments <- c("S", "I", "R")
params <- c("beta", "mu")
extract_rate_fcns(rates, compartments, params)
}

