% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_keys.R
\name{generate_keys}
\alias{generate_keys}
\title{Generate the keys for indexing into the .irm and .eigen environments.}
\usage{
generate_keys(epimodel, inds = NULL)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{inds}{optional vector with row numbers for which keys are desired.}
}
\value{
character vector of keys
}
\description{
Generate the keys for indexing into the .irm and .eigen environments.
}

