% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval_rates.R
\name{get_rate_mat}
\alias{get_rate_mat}
\title{Evaluate the rate functions at the current state and the parameter values,
returning a matrix of rates for use in the \code{simulate_epimodel} function.}
\usage{
get_rate_mat(rate_fcns, state, flow, params)
}
\arguments{
\item{rate_fcns}{list of rate functions created by extract_rate_fcns}

\item{state}{current configuration of the population state vector}

\item{flow}{numeric matrix with reactions as rows and where each column in a
row has element 1 to indicate an entry to that compartment, -1 to indicate
an exit, and 0 for no change in the size of the compartment on the subject
level. Columns should have names corresponding to the state labels.}

\item{params}{character vector of process parameters with element names
corresponding exactly to names of parameters used in the \code{rates}
arguments.}
}
\value{
matrix of rates
}
\description{
Evaluate the rate functions at the current state and the parameter values,
returning a matrix of rates for use in the \code{simulate_epimodel} function.
}

