% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpms_to_build.R
\name{get_tpms_to_build}
\alias{get_tpms_to_build}
\title{Get the indices for the transition probability matrices to be rebuilt}
\usage{
get_tpms_to_build(epimodel, subject = NULL)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects.}

\item{subject_ID}{string ID for subject to be removed}
}
\value{
vector of indices
}
\description{
Get the indices for the transition probability matrices to be rebuilt
}

