% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getObsTimeInds}
\alias{getObsTimeInds}
\title{Get the indices of observation times in the population level bookkeeping mtx.}
\usage{
getObsTimeInds(pop_mat, obstimes)
}
\arguments{
\item{pop_mat}{population level bookkeeping matrix}

\item{obstimes}{vector of observation times}
}
\value{
indices of observation times
}
\description{
Get the indices of observation times in the population level bookkeeping mtx.
}

