% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_template.R
\name{gibbs_template}
\alias{gibbs_template}
\title{Template for Gibbs kernel}
\usage{
gibbs_template(epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
Template for Gibbs updates - example given for SIR model
}
\description{
Template for writing a MH transition kernel, example based on SIR model with
Gibbs updates for the sampling parameter.
}

