% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_augmentation.R
\name{init_augmentation}
\alias{init_augmentation}
\title{Generate an initial configuration that is concordant with the data, updating
the configuration and observation matrices.}
\usage{
init_augmentation(epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
initialized epimodel object
}
\description{
Generate an initial configuration that is concordant with the data, updating
the configuration and observation matrices.
}

