% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_bookkeeping.R
\name{init_bookkeeping}
\alias{init_bookkeeping}
\title{Initializes a list of bookkeeping matrices for the observations, population
level trajectory, and subject level trajectory.}
\usage{
init_bookkeeping(dat = TRUE, pop = TRUE, subj = TRUE, obstimes, meas_vars,
  init_state, tmax)
}
\arguments{
\item{dat}{logical indicator for whether to include the data matrix in the
bookkeeping list.}

\item{pop}{logical indicator for whether to include the population level
trajectory matrix in the bookkeeping list.}

\item{subj}{logical indicator for whether to include the subject level
trajectory matrix in the bookkeeping list.}

\item{obstimes}{numeric vector of observation times with first element the
time at which the process is initialized, and last element the final time
of observation. The user may specify \code{Inf} as an option, in which case
the epidemic is completely observed.}

\item{meas_vars}{a character vector specifying which compartments are
measured}

\item{init_state}{numeric vector of initial states of compartments the system
with named elements corresponding exactly to the names of compartments used
in the \code{rates} argument.}

\item{tmax}{time until which to simulate the process, possibly infinity.}
}
\value{
list of matrices for data, population trajectory, and subject level trajectory
}
\description{
Initializes a list of bookkeeping matrices for the observations, population
level trajectory, and subject level trajectory.
}

