% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_data_mat.R
\name{init_data_mat}
\alias{init_data_mat}
\title{Initializes a bookkeeping object for the data}
\usage{
init_data_mat(obstimes, meas_vars)
}
\arguments{
\item{obstimes}{numeric vector of observation times with first element the
time at which the process is initialized, and last element the final time
of observation. The user may specify \code{Inf} as an option, in which case
the epidemic is completely observed.}

\item{meas_vars}{a character vector specifying which compartments are
measured}
}
\value{
Bookkeeping object for observed and true counts at observation times.
}
\description{
Initializes a bookkeeping object for the data
}
\examples{
obstimes <- seq(0, 5, by = 0.5)
meas_vars <- "I"
init_data_obj(obstimes, meas_vars)
}

