% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_epimodel.R
\name{init_epimodel}
\alias{init_epimodel}
\title{Initialize epimodel bookkeeping object.}
\usage{
init_epimodel(states, params, rates, flow, dat = NULL, time_var = NULL,
  obstimes = NULL, popsize = NULL, pop_mat = NULL, subj_mat = NULL,
  obs_mat = NULL, meas_vars = NULL, r_meas_process = NULL,
  d_meas_process = NULL, covar = NULL, tcovar = NULL, rprior = NULL,
  dprior = NULL, to_estimation_scale = NULL, from_estimation_scale = NULL,
  init_state = NULL, initialization_fcn = NULL)
}
\arguments{
\item{states}{character vector with names of compartments corresponding
exactly to the names of compartments used in the \code{*_rates} arguments.}

\item{params}{character vector of process parameters with element names
corresponding exactly to names of parameters used in the \code{rates}
arguments.}

\item{rates}{list of functions or a character vector with strings specifying
the flow rates between compartments with compartments corresponding to
names given by \code{states}.}

\item{flow}{numeric matrix with reactions as columns and where each row in a
column indicates the change in the size of the compartment on the subject
level.}

\item{dat}{matrix of dimension \code{number of observation times \emph{x}
number of measured compartments}. \code{dat} must have one column with
observation times (numeric and strictly increasing), whose the name of
which is given by the \code{times} variable. The matrix must be sorted in
ascending time order.}

\item{time_var}{string indicating the name of the variable coding observation
times in \code{dat}.}

\item{popsize}{size of the population.}

\item{pop_mat}{population-level bookkeeping matrix for compartment counts at
times of state transition.}

\item{subj_mat}{subject-level bookkeeping matrix for individual trajectories.}

\item{obs_mat}{population-level bookkeeping matrix for compartment counts at
observation times.}

\item{meas_vars}{character vector specifying which compartments are measured.}

\item{r_meas_process}{function to simulate from the measurement process, with
named arguments \code{proc_state}, \code{meas_vars}, and \code{params},
which are the current state of the process (given as a named character
vector with element names corresponding exactly to the names of
compartments), a vector with the names of the states to be measured, and a
named vector of process parameters. The function should return noisy
measurements of the process in a vector with named elements corresponding
to each of the measured compartments specified in \code{meas_vars}.}

\item{d_meas_process}{function to evaluate the density of the measurement
process with arguments and output specified as in \code{r_meas_process}.}

\item{covar}{optional numeric matrix of time varying covariates (currently
not working).}

\item{tcovar}{numeric vector of times for time-varying covariates (currently
not working).}

\item{rprior}{list of functions for simulating from the prior distributions
of model parameters. List element names should correspond to exactly to the
parameter names given in \code{params}.}

\item{dprior}{list of functions for evaluating the prior densities of model
parameters. List element names should correspond to exactly to the
parameter names given in \code{params}.}

\item{to_estimation_scale}{list of functions for transforming model
parameters to the scale on which new parameters should be proposed. List
element names should correspond to exactly to the parameter names given in
\code{params}.}

\item{from_estimation_scale}{list of functions for transforming model
parameters from the a scale on which new parameters are proposed to the
scale used to evaluate the process likelihood. List element names should
correspond to exactly to the parameter names given in \code{params}.}

\item{init_state}{numeric vector of initial states of compartments the system
with named elements corresponding exactly to the names of compartments used
in the \code{rates} argument.}

\item{initialization_function}{optional function requiring no inputs that
  can be provided in place of init_state to simulate the initial state of the
  system. The function must output a named vector of the same form as
  \code{init_state}.}
}
\value{
list containing bookkeeping objects and model configuration objects.
}
\description{
Initializes a bookkeeping list containing the population level and subject
level bookkeeping objects, measurement, and epidemic process settings. At a
minimum, the epimodel object must be initialized with the states, params,
rates, flow, and either the vector of observation times or a matrix with
data.
}

