% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_irm.R
\name{init_irm}
\alias{init_irm}
\title{Initialize an arrays for storing matrices of rates, eigenvectors, and inverse matrices for eigenvectors, along with a matrix of eigenvalues.}
\usage{
init_irm(epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
epimodel object with initialized objects
}
\description{
Initialize an arrays for storing matrices of rates, eigenvectors, and inverse matrices for eigenvectors, along with a matrix of eigenvalues.
}

