% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_obs_mat.R
\name{init_obs_mat}
\alias{init_obs_mat}
\title{Initializes a bookkeeping object for the observed and augmented data.}
\usage{
init_obs_mat(epimodel)
}
\arguments{
\item{epimodel}{bookkeeping object.}
}
\value{
Bookkeeping object for observed and true counts at observation times.
}
\description{
Initializes a bookkeeping object for the observed and augmented data.
}
