% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_pop_mat.R
\name{init_pop_mat}
\alias{init_pop_mat}
\title{initializes the bookkeeping object for population level counts}
\usage{
init_pop_mat(init_state, tmax)
}
\arguments{
\item{init_state}{numeric vector of initial states of compartments the system
with named elements corresponding exactly to the names of compartments used
in the \code{rates} argument.}

\item{tmax}{time until which to simulate the process, possibly infinity.}
}
\value{
initialized matrix with columns to store event times and counts of
  individuals in each compartment. The first and last rows of the matrix
  always correspond to time 0 and tmax. The state at tmax is initialized to
  init_state.
}
\description{
initializes the bookkeeping object for population level counts
}
\examples{
init_state <- c(S = 45, I = 5, R = 0)
tmax <- 5
init_pop_traj(init_state, tmax)
}

