% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_rate_map.R
\name{init_rate_map}
\alias{init_rate_map}
\title{Insert a lookup table to map subjects at risk for each transition to the appropriate rate code label}
\usage{
init_rate_map(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects. If not supplied, the
list will be generated.}
}
\value{
character vector for mapping state codes to compartment names.
}
\description{
Insert a lookup table to map subjects at risk for each transition to the appropriate rate code label
}

