% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_results.R
\name{init_results}
\alias{init_results}
\title{Initialize list for storing results}
\usage{
init_results(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects.}

\item{sim_settings}{bookkeeping list of simulation settings}
}
\value{
list containing the matrix of posterior parameter draws,
  configurations, log-likelihoods, total simulation time, and seed.
}
\description{
Initialize list for storing results
}

