% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_state_lookup.R
\name{init_state_lookup}
\alias{init_state_lookup}
\title{Insert a lookup table to connect state indicators to state labels.}
\usage{
init_state_lookup(states)
}
\value{
data.frame for lookup table for state names and state codes.
}
\description{
Insert a lookup table to connect state indicators to state labels.
}

