% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_trajectories.R
\name{init_trajectories}
\alias{init_trajectories}
\title{Initialize a set of trajectories if an initial set of paths not provided.}
\usage{
init_trajectories(epimodel)
}
\value{
epimodel object with initialized configuration matrix
}
\description{
Initialize a set of trajectories if an initial set of paths not provided.
}
