% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_row.R
\name{insert_row}
\alias{insert_row}
\title{Generic function to insert a row into a matrix.}
\usage{
insert_row(mat, row_num, vec)
}
\arguments{
\item{mat}{matrix into which row should be inserted}

\item{row_num}{row number for insertion; row currently at that number is pushed down.}

\item{vec}{vector to be inserted}
}
\value{
matrix with the inserted row.
}
\description{
Generic function to insert a row into a matrix.
}
\examples{
x <- matrix(1:9, nrow = 3)
insert_row(x, 2, c(5, 27, 87))

}

