% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert_trajectory.R
\name{insert_trajectory}
\alias{insert_trajectory}
\title{Insert a newly sampled subject-level trajectory so its contribution is
reflected in the configuration matrix.}
\usage{
insert_trajectory(epimodel, subject, subj_ID)
}
\arguments{
\item{subj_ID}{character string for subject ID code}
}
\value{
updated configuration matrix in the epimodel environment
}
\description{
This function also updates several bookkeeping objects - .ind_final_config
}

