% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insertPath}
\alias{insertPath}
\title{Insert subject transitions into the population-level and subject level 
bookkeeping matrices.}
\usage{
insertPath(path, subject, pop_mat, config_mat, ind)
}
\arguments{
\item{path}{matrix of transitions to be inserted}

\item{subject}{subject ID}

\item{pop_mat}{population bookkeeping matrix in epimodel list}

\item{config_mat}{subject-level bookkeeping matrix}

\item{subj_row_ind}{index stating where to begin inserting the transitions}
}
\value{
updated bookeeping objects
}
\description{
Insert subject transitions into the population-level and subject level 
bookkeeping matrices.
}

