% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irm_decomp.R
\name{irm_decomp}
\alias{irm_decomp}
\title{Compute the eigen decomposition for a rate matrix, and either instatiate it 
or update an existing decomposition.}
\usage{
irm_decomp(irm, env = NULL, key = NULL, update = TRUE)
}
\arguments{
\item{irm}{rate matrix}

\item{env}{environment into which to instatiate the rate matrix}

\item{key}{the key for the hash index}

\item{update}{TRUE/FALSE for whether to update an existing rate matrix or to
instatiate a new one}
}
\value{
list containing the unordered eigenvalues, the matrix of
  eigenvectors, and the inverse of the eigenvector matrix
}
\description{
Compute the eigen decomposition for a rate matrix, and either instatiate it 
or update an existing decomposition.
}

