% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_row.R
\name{match_row}
\alias{match_row}
\title{Returns either a logical vector of indices for which row(s) in a table match 
a given vector or rows in a matrix.}
\usage{
match_row(x, table, return_value)
}
\arguments{
\item{x}{vector or matrix}

\item{table}{table in which to search for matches}

\item{return_value}{one of "index", "logical", or "vector" to return the index
of the matching row (possibly of length 0), a logical value indicating
whether there is a matching row, or a vector of logicals indicating a match
or not.}
}
\value{
vector of logicals or indices
}
\description{
Returns either a logical vector of indices for which row(s) in a table match 
a given vector or rows in a matrix.
}

