% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_template.R
\name{mh_template}
\alias{mh_template}
\title{Template for Metropolis-Hastings kernel}
\usage{
mh_template(epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
template for Metropolis-Hastings updates - example given for SIR model
}
\description{
Template for writing a MH transition kernel, example based on SIR model with
MH proposals for infectivity and recovery parameters on log transformed scale.
}
