% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_posterior.R
\name{plot_latent_posterior}
\alias{plot_latent_posterior}
\title{Plot the posterior distribution of the latent paths for measured states.}
\usage{
plot_latent_posterior(epimodel, states, times)
}
\arguments{
\item{epimodel}{fitted epimodel object}

\item{states}{character vector of states for which to plot the posterior}

\item{times}{sequence of times at which to compute the posterior}
}
\value{
plot of the posterior distribution of the latent paths
}
\description{
Plot the posterior distribution of the latent paths for measured states.
}

