% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{populationLikelihood}
\alias{populationLikelihood}
\title{Get the irm keys for the compartment counts in a population level
bookkeeping matrix}
\usage{
populationLikelihood(pop_mat, irm, initdist, initdist_param_inds,
  flow_inds, keys, inds, loglik)
}
\arguments{
\item{pop_mat}{population level bookkeeping matrix}

\item{irm}{array of rate matrices}

\item{initdist}{vector of initial state probabilities}

\item{initdist_param_inds}{vector of indices for admissible initial states}

\item{flow_inds}{matrix of indices for locations in IRM for each event}

\item{keys}{vector of irm keys}

\item{inds}{indices relating to event times, along with t0 and tmax}

\item{loglik}{boolean indicating whether to return the likelihood or
log-likelihood}
}
\value{
population level likelihood or log-likelihood
}
\description{
Get the irm keys for the compartment counts in a population level
bookkeeping matrix
}
