% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_epimodel.R
\name{prepare_epimodel}
\alias{prepare_epimodel}
\title{Prepares an epimodel environment.}
\usage{
prepare_epimodel(epimodel)
}
\arguments{
\item{epimodel}{initialized epimodel list}
}
\value{
epimodel environment
}
\description{
Prepares an epimodel environment.
}
