% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinsert_path.R
\name{reinsert_path}
\alias{reinsert_path}
\title{Re-insert the current path back into the configuration matrix.}
\usage{
reinsert_path(epimodel, subject)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects.}
}
\value{
updated configuration matrix
}
\description{
Note that this gets called after the proposed path has been removed.
Therefore, .ind_final_config is already updated.
}
