% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_trajectory.R
\name{remove_trajectory}
\alias{remove_trajectory}
\title{Remove the contribution of a subject from the compartment counts in the
\code{config_mat} and \code{obs_mat} objects.}
\usage{
remove_trajectory(epimodel, subject)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{subject}{ID for subject to be removed}
}
\value{
\code{config_mat} and \code{obs_mat} objects within the .epimodel
  environment.
}
\description{
Remove the contribution of a subject from the compartment counts in the
\code{config_mat} and \code{obs_mat} objects.
}

