% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeSubjPath}
\alias{removeSubjPath}
\title{Add or remove the subject contribution from the compartment counts}
\usage{
removeSubjPath(pop_mat, config_mat, subj_inds)
}
\arguments{
\item{pop_mat}{population bookkeeping matrix in epimodel list}

\item{config_mat}{subject level configuration bookkeeping matrix}

\item{subj_inds}{indices for rows relating to the subject's events}
}
\value{
updated compartment counts
}
\description{
Add or remove the subject contribution from the compartment counts
}

