% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resolveSubjContrib}
\alias{resolveSubjContrib}
\title{Add or remove the subject contribution from the compartment counts}
\usage{
resolveSubjContrib(pop_mat, ind_final_config, subj_path, insertion)
}
\arguments{
\item{pop_mat}{population bookkeeping matrix in epimodel list}

\item{subj_path}{column inds given by the subject path}

\item{insertion}{Logical indicating whether the resolution should be the
insertion or the removal of a path}

\item{row_inds}{vector of row indices, 1:epimodel$ind_final_config}
}
\value{
updated compartment counts
}
\description{
Add or remove the subject contribution from the compartment counts
}
