% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieveKeys}
\alias{retrieveKeys}
\title{Get the irm keys for the compartment counts in a population level
bookkeeping matrix}
\usage{
retrieveKeys(inds, irm_lookup, pop_mat, index_state_num)
}
\arguments{
\item{inds}{vector of indices, generally 1:epimodel$ind_final_config}

\item{irm_lookup}{lookup matrix relating configurations to irm keys}

\item{pop_mat}{population bookkeeping matrix in epimodel list}

\item{keys}{vector of keys to be modified in place}

\item{append_missing}{indicator to be modified in place to indicate whether 
any keys are missing}

\item{index_state_number}{vector indicating which states are index states}
}
\value{
Vector of keys to index into an IRM array.
}
\description{
Get the irm keys for the compartment counts in a population level
bookkeeping matrix
}

