% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{retrieveSubjPath}
\alias{retrieveSubjPath}
\title{Insert subject transitions into the population-level and subject level 
bookkeeping matrices.}
\usage{
retrieveSubjPath(subj_path, subject, pop_mat, init_config, ind_final_config,
  flow_inds)
}
\arguments{
\item{subj_path}{subject path vector}

\item{subject}{ID for subject}

\item{pop_mat}{population level bookkeeping matrix}

\item{init_config}{initial configuration vector}

\item{ind_final_config}{index for final configuration}

\item{flow_inds}{matrix of flow indices}
}
\value{
extended path vector
}
\description{
Insert subject transitions into the population-level and subject level 
bookkeeping matrices.
}
