% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_at_obs_times.R
\name{sample_at_obs_times}
\alias{sample_at_obs_times}
\title{Sample the infection status for the specified subject at observation times.}
\usage{
sample_at_obs_times(path, epimodel)
}
\arguments{
\item{path}{vector giving a subject path.}

\item{epimodel}{list of bookkeeping and model objects.}
}
\value{
updated path vector
}
\description{
Sample the infection status for the specified subject at observation times.
}
