% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_DT_skeleton.R
\name{sample_DT_skeleton}
\alias{sample_DT_skeleton}
\title{Sample the state at event times from the discete time skeleton.}
\usage{
sample_DT_skeleton(config_mat, epimodel, subject, init_ind, final_ind)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{subject}{subject ID, i.e. column index in config_mat}

\item{init_ind,final_ind}{index in the configuration matrix of the left/right
  endpoints of the observation interval}
}
\value{
updated configuration matrix
}
\description{
This function only gets called when there is at least one transition event in
an inter-observation interval
}

