% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_forward.R
\name{sample_forward}
\alias{sample_forward}
\title{Forward path sampling.}
\usage{
sample_forward(path, epimodel, subject, init_time, final_time, init_state,
  final_state, irm_key)
}
\arguments{
\item{path}{matrix for storing path, possibly with first transition completed}

\item{irm_key}{}
}
\value{
updated bookkeeping matrix with a valid path (no update if constant
  path)
}
\description{
Forward path sampling.
}

