% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_path_in_interval.R
\name{sample_path_in_interval}
\alias{sample_path_in_interval}
\title{Draw a path within an interval from an endpoint conditioned homogeneous CTMC.}
\usage{
sample_path_in_interval(epimodel, subject, interval)
}
\arguments{
\item{subject}{character string for subject ID code}

\item{interval}{index for the interval within which path is to be drawn}
}
\value{
matrix of transitions for path or NULL
}
\description{
Draw a path within an interval from an endpoint conditioned homogeneous CTMC.
}

