% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_epimodel.R
\name{simulate_epimodel}
\alias{simulate_epimodel}
\title{Simulates a stochastic epidemic model.}
\usage{
simulate_epimodel(epimodel, init_state = NULL, initialization_fcn = NULL,
  return_config = TRUE, trim = TRUE)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects. If not supplied, the
list will be generated.}

\item{return_config}{option specifying whether to return a matrix of
subject-level trajectories. Defaults to \code{TRUE}.}

\item{trim}{option specifying whether to trim the simulated data if the
  epidemic ends before the final observation time. Defaults to \code{TRUE}.}

\item{initialization_function}{optional function requiring no inputs that can
be provided in place of init_state to simulate the initial state of the
system. The function must output a named vector of the same form as
\code{init_state}.}
}
\value{
Data frame with columns for the observation times, and draws from the
  measurement process at observation times. Optionally, also return a data
  frame with columns for event times and states of the epidemic compartments
  at each event time.
}
\description{
Simulate a realization from a stochastic epidemic model along with a dataset
resulting from a specified measurement process. The (latent) population
trajectory for the epidemic is simulated using Gillespie's first reaction
method (Gillespie 1976, 1977), with the trajectory optionally being returned
in the output. The dataset that is returned consists of observations of the
epidemic that are distributed according to a measurement process, provided as
a function by the user. The dynamics of the system are characterized by a
named vector of flow rates between compartments and a numeric matrix that
specifies the changes to each compartment due to each type of event. The user
also specifies a named vector of process parameters, a named vector of
initial states, and a set of times at which the epidemic process is sampled.
The syntax for calling the function is deliberately similar to that used in
the 'GillespieSSA' package so as to facilitate ease of use for users familiar
with that package.
}
\references{
{Gillespie DT (1976). “A general method for numerically
  simulating the stochastic time evolution of coupled chemical reactions.”
  Journal of Computational Physics, *22*(4), pp. 403-434.

  Gillespie DT (1977). “Exact stochastic simulation of coupled chemical
  reactions.” The Journal of Physical Chemistry, *81*(25), pp. 2340- 2361.}
}

