% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SIR.R
\name{simulate_SIR}
\alias{simulate_SIR}
\title{Simulate an epidemic with SIR dynamics and draw binomial samples.}
\usage{
simulate_SIR(obstimes, params, popsize, init_config = NULL,
  trim = TRUE)
}
\arguments{
\item{obstimes}{vector of observation times}

\item{params}{vector of parameters, must be in the following order: beta, mu,
rho, S0, I0, R0}

\item{popsize}{population size}

\item{init_config}{compartment counts at t0.}

\item{trim}{logical for whether to trim the observations after the epidemic 
has died off}
}
\value{
list with a population bookkeeping matrix and an observation matrix
}
\description{
Simulate an epidemic with SIR dynamics and draw binomial samples.
}
\examples{
obstimes <- 0:10
params <- c(0.04, 1, 0.5, 0.9, 0.03, 0.07)
popsize <- 100

epidemic <- simulate_SIR(obstimes, params, popsize)
}
