% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_probabilities.R
\name{subject_probabilities}
\alias{subject_probabilities}
\title{Build a vector of probabilities according to which subjects should be
resampled.}
\usage{
subject_probabilities(epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
vector of probabilities according to which subject IDs are sampled
}
\description{
Build a vector of probabilities according to which subjects should be
resampled.
}
