% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subjectLikelihood}
\alias{subjectLikelihood}
\title{Get the irm keys for the compartment counts in a population level
bookkeeping matrix}
\usage{
subjectLikelihood(subject, pop_mat, config_mat, irm_array, initdist, keys, inds,
  loglik)
}
\arguments{
\item{subject}{}

\item{pop_mat}{population level bookkeeping matrix}

\item{config_mat}{subject_level bookkeeping matrix for configurations}

\item{irm_array}{array of rate matrices}

\item{keys}{vector of irm keys}

\item{inds}{indices relating to event times, along with t0 and tmax}

\item{loglik}{boolean indicating whether to return the likelihood or
    log-likelihood}
}
\value{
subject level likelihood or log-likelihood
}
\description{
Get the irm keys for the compartment counts in a population level
bookkeeping matrix
}

