% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_template.R
\name{templates}
\alias{templates}
\title{Generates a template script for specifying a transition kernel or functions 
to and from the estimation scale.}
\usage{
templates(type)
}
\arguments{
\item{type}{either "mh" for a Metropolis-Hastings kernel, "gibbs" for a Gibbs
kernel, or "est_scale" for a transformation function from the model scale
to the parameter estimation scale.}
}
\value{
template
}
\description{
Generates a template script for specifying a transition kernel or functions 
to and from the estimation scale.
}

