% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_scale_template.R
\name{to_est_scale_template}
\alias{to_est_scale_template}
\title{Template for function for transformint to the estimation scale}
\usage{
to_est_scale_template(params, epimodel)
}
\arguments{
\item{epimodel}{}
}
\value{
template for transformation functions.
}
\description{
Template for writing a MH transition kernel, example based on SIR model, 
where proposals for the infectivity and recovery parameters are made in terms
of log(R_0), and log(mu). See package vignette for additional details. The
function for transforming from the estimation scale back to the model scale
should be similarly specified.
}

