% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpmProdSeqs}
\alias{tpmProdSeqs}
\title{Update eigen values, vectors, and inverse matrices for irms}
\usage{
tpmProdSeqs(tpm_prods, tpms, obs_time_inds)
}
\arguments{
\item{tpm_prods}{array of transition probability matrix products}

\item{tpms}{array of transition probability matrices}

\item{pop_mat}{population level bookkeeping matrix}

\item{obstimes}{vector of observation times}
}
\value{
Updated eigenvalues, eigenvectors, and inverse matrices
}
\description{
Update eigen values, vectors, and inverse matrices for irms
}

