% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tpms.R
\name{update_tpms}
\alias{update_tpms}
\title{Update the list of tpms and tpm products to reflect the removal or insertion 
of a subject-level trajectory.}
\usage{
update_tpms(epimodel, subject, direction)
}
\arguments{
\item{direction}{"removal" or "insertion"}
}
\value{
updated lists of tpms within the .epimodel environment
}
\description{
Gets called before \code{remove_trajectory} in the removal direction, and
after \code{draw_trajec} in the insertion direction.
}
