% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/append_missing.R
\name{append_missing}
\alias{append_missing}
\title{Append missing objects to the irm arrays and eigen decomposition objects.
Update irm_key_lookup matrix.}
\usage{
append_missing(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}
}
\value{
epimodel with complete irm and eigen decomposition arrays.
}
\description{
Append missing objects to the irm arrays and eigen decomposition objects.
Update irm_key_lookup matrix.
}

