% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_emission_mat.R
\name{build_emission_mat}
\alias{build_emission_mat}
\title{Updates the matrix of emission probabilities}
\usage{
build_emission_mat(emission_mat, epimodel)
}
\arguments{
\item{emission_mat}{matrix of emission probabilities to be updated}

\item{epimodel}{}
}
\value{
updated matrix of emission probabilities
}
\description{
Updates the matrix of emission probabilities
}
