% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_initdist.R
\name{build_initdist}
\alias{build_initdist}
\title{Construct the vector with initial state probabilities for a single subject at
time t0.}
\usage{
build_initdist(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects.}
}
\value{
vector of initial state probabilities for a single subject
}
\description{
Construct the vector with initial state probabilities for a single subject at
time t0.
}
