% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rate_mat.R
\name{build_rate_mat}
\alias{build_rate_mat}
\title{Construct a single rate matrix from a vector of rates and a flow matrix.}
\usage{
build_rate_mat(rates, state, params, flow, env = NULL, key = NULL,
  update = TRUE)
}
\arguments{
\item{rates}{vector containing values of evaluated rates}

\item{params}{numeric vector of process parameters with element names 
corresponding exactly to names of parameters used in the \code{rates} 
arguments. The params vector must contain a number of parameters equal to 
the number of states for the categorical distribution for the state of a 
subject at t0. The names of these parameters must end in 0 - e.g. for the 
SIR model: S0, I0, R0. A value of zero for one of the initial distribution 
parameters will be interpretted as a model assumption that subjects cannot 
occupy that state at time t0.}

\item{flow}{numeric matrix of dimension \code{number of transitions \emph{x} 
number of compartments}. Each row corresponds to a possible transition, and
each column in a row has element 1 to indicate an entry to that 
compartment, -1 to indicate an exit, and 0 for no change in the size of the
compartment on the subject level.}

\item{env}{environment into which to instatiate the rate matrix}

\item{key}{the key for the hash index}

\item{update}{TRUE/FALSE for whether to update an existing rate matrix or to
instatiate a new one}
}
\value{
rate matrix of dimension \code{num_states x num_states}
}
\description{
Construct a single rate matrix from a vector of rates and a flow matrix.
}

