% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildEigenArray}
\alias{buildEigenArray}
\title{Update eigen values, vectors, and inverse matrices for irms}
\usage{
buildEigenArray(eigenvals, eigenvecs, inversevecs, irm_array)
}
\arguments{
\item{eigenvals}{matrix of eigenvalues with columns corresponding to irms}

\item{eigenvecs}{array of eigenvectors with third dimension corresponding to
irms}

\item{inversevecs}{array of inverses of matrices of eigenvectors}

\item{irm_array}{array of rate matrices}
}
\value{
Updated eigenvalues, eigenvectors, and inverse matrices
}
\description{
Update eigen values, vectors, and inverse matrices for irms
}

