% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_pop_likelihood.R
\name{calc_pop_likelihood}
\alias{calc_pop_likelihood}
\title{Calculate likelihood for the population level trajectory}
\usage{
calc_pop_likelihood(epimodel, log = FALSE)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{log}{TRUE/FALSE for whether to return the log-likelihood or likelihood.}
}
\value{
vector with log-likelihood of the population level trajectory and
  measurement process log-likelihood.
}
\description{
Calculate likelihood for the population level trajectory
}

