% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rate_fcns.R
\name{extract_rate_fcns}
\alias{extract_rate_fcns}
\title{Extracts a list of functions for computing flow rates.}
\usage{
extract_rate_fcns(rates, states, param_names)
}
\arguments{
\item{states}{vector of compartment names}

\item{param_names}{vector of parameter names}
}
\value{
Returns a list of functions to compute rates.
}
\description{
Extracts a list of functions for computing flow rates.
}
\examples{
rates <- c("beta * I", "mu")
states <- c("S", "I", "R")
param_names <- c("beta", "mu")
extract_rate_fcns(rates, states, param_names)      
       
}
