% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_epimodel.R
\name{fit_epimodel}
\alias{fit_epimodel}
\title{Main wrapper to fit a stochastic epimodel using the Bayesian data
augmentation algorithm.}
\usage{
fit_epimodel(epimodel, monitor = FALSE)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{monitor}{TRUE/FALSE for whether to print the iteration number and
  produce a log-likelihood plot every time a trajectory is saved.}
}
\value{
list containing the epimodel object and a results list.
}
\description{
Main wrapper to fit a stochastic epimodel using the Bayesian data
augmentation algorithm.
}

