% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_epimodel.R
\name{fit_epimodel}
\alias{fit_epimodel}
\title{Main wrapper to fit a stochastic epimodel using the Bayesian data
augmentation algorithm.}
\usage{
fit_epimodel(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects. If not supplied, the
list will be generated.}
}
\value{
list containing the epimodel object and a results list.
}
\description{
Main wrapper to fit a stochastic epimodel using the Bayesian data
augmentation algorithm.
}

