% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_keys.R
\name{generate_keys}
\alias{generate_keys}
\title{Generate the keys for indexing into the .irm and .eigen environments.}
\usage{
generate_keys(epimodel, inds = NULL, lookup = TRUE)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects.}

\item{inds}{optional vector with row numbers for which keys are desired.}

\item{lookup}{logical indicating whether to look up the keys in the
  .irm_key_lookup object}
}
\value{
character vector of keys
}
\description{
Generate the keys for indexing into the .irm and .eigen environments.
}

