% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_rate_map.R
\name{init_rate_map}
\alias{init_rate_map}
\title{Construct an index matrix identifying with which states each rate varies.}
\usage{
init_rate_map(rates, states)
}
\arguments{
\item{rates}{character vector with strings specifying the flow rates between
compartments. Each variable in a rate must correspond to the name of one of
the \code{states}, \code{params}, or \code{covars} (covariates not yet
implemented).}

\item{states}{character vector with names of compartments corresponding
exactly to the names of compartments used in the \code{*_rates} arguments.}
}
\value{
binary index matrix indicating which changes in each compartment
  affect each rate
}
\description{
Called internally before rates are parsed when an epimodel object is 
initialized.
}

