% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_settings.R
\name{init_settings}
\alias{init_settings}
\title{Initialize a bookkeeping list of MCMC settings.}
\usage{
init_settings(epimodel, niter, burnin, params_every = 1,
  configs_every = floor(niter/100), kernel, cov_mtx = NULL,
  configs_to_redraw, to_estimation_scale = NULL,
  from_estimation_scale = NULL)
}
\arguments{
\item{niter}{number of iterations/parameter updates for which to run the
sampler.}

\item{burnin}{length of burn-in run in terms of number of parameter updates.}

\item{params_every}{thin parameter updates by saving every k-th draw.}

\item{configs_every}{thin configurations by saving the current configuration
after every k-th parameter update.}

\item{kernel}{transition kernel for parameters, provided as a list of
transition functions. Each function should use objects that are available
within an epimodel list or that are supplied as arguments to
init_simulation.}

\item{cov_mtx}{covariance matrix for parameters to be used in M-H updates.}

\item{configs_to_redraw}{number of subject level trajectories to sample
between parameter updates.}

\item{to_estimation_scale}{list of functions for transforming model
parameters to the scale on which new parameters should be proposed. List
element names should correspond to exactly to the parameter names given in
\code{params}.}

\item{from_estimation_scale}{list of functions for transforming model
parameters from the a scale on which new parameters are proposed to the
scale used to evaluate the process likelihood. List element names should
correspond to exactly to the parameter names given in \code{params}.}

\item{priors}{list of functions to evaluate prior distributions.}

\item{parallelize}{option to parallelize computation of transition
  probability and rate matrices. Defaults to FALSE.}
}
\value{
bookkeeping list
}
\description{
Initialize a bookkeeping list of MCMC settings.
}
\examples{
# Transition kernel for the SIR model - block MH updates for beta and mu,
# Gibbs updates for rho and p0. cov_mtx is supplied as an additional argument
# to init_simulation. This function uses the mvrnorm function from the MASS
# package.

# functions to go to and from the estimation scale
}

