% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_state_lookup.R
\name{init_state_lookup}
\alias{init_state_lookup}
\title{Insert a lookup table to map subjects at risk for each transition to the appropriate rate code label}
\usage{
init_state_lookup(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}
}
\value{
character vector for mapping state codes to compartment names.
}
\description{
Insert a lookup table to map subjects at risk for each transition to the appropriate rate code label
}

