% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_subj_mat.R
\name{init_subj_mat}
\alias{init_subj_mat}
\title{Initializes a matrix of subject level trajectories.}
\usage{
init_subj_mat(epimodel, init_state)
}
\arguments{
\item{epimodel}{bookkeeping object}

\item{init_state}{numeric vector of initial states of compartments the system
with named elements corresponding exactly to the names of compartments used
in the \code{rates} argument.}
}
\value{
matrix of subject level trajectories, where each subject has two
  rows, one for time zero and one for tmax.
}
\description{
Initializes a matrix of subject level trajectories.
}

