% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_trajectory.R
\name{insert_trajectory}
\alias{insert_trajectory}
\title{Insert a newly sampled subject-level trajectory so its contribution is 
reflected in the configuration matrix.}
\usage{
insert_trajectory(epimodel, subject, reinsertion)
}
\arguments{
\item{subject}{character string for subject ID code}

\item{reinsertion}{TRUE/FALSE for whether the trajectory is to be be
reinserted following a MH rejection}
}
\value{
updated epimodel object
}
\description{
This function also updates several bookkeeping objects - .ind_final_config
}

