% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{joinCubes}
\alias{joinCubes}
\title{Join two armadillo cubes}
\usage{
joinCubes(firstcube, secondcube)
}
\arguments{
\item{cube2}{}
}
\value{
new array joining cubes
}
\description{
Join two armadillo cubes
}

