% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MH_accept_reject.R
\name{MH_accept_reject}
\alias{MH_accept_reject}
\title{Accept or reject a proposed subject-level trajectory via MH and update the
appropriate objects in the epimodel environment}
\usage{
MH_accept_reject(epimodel, subject, subj_ID)
}
\value{
updated epimodel environment
}
\description{
Accept or reject a proposed subject-level trajectory via MH and update the
appropriate objects in the epimodel environment
}

