% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_posterior.R
\name{plot_latent_posterior}
\alias{plot_latent_posterior}
\title{Plot the posterior distribution of the latent paths for measured states.}
\usage{
plot_latent_posterior(epimodel, states, times, cm = "med",
  proportion = FALSE, truth = NULL, true_popsize = NULL, dat = NULL)
}
\arguments{
\item{epimodel}{fitted epimodel object}

\item{states}{character vector of states for which to plot the posterior}

\item{times}{sequence of times at which to compute the posterior}

\item{cm}{"med" if posterior median, 2.5 and 97.5 quantiles are desired,
otherwise "mn" if posterior mean, and 95 percent CI are desired}

\item{proportion}{if true, plot the proportion instead of the counts}

\item{truth}{true latent trajectory}

\item{true_popsize}{true population size, defaults to the population size 
specified in the epimodel object unless otherwise provided.}

\item{dat}{data matrix in the epimodel object}
}
\value{
plot of the posterior distribution of the latent paths
}
\description{
Plot the posterior distribution of the latent paths for measured states.
}

