% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_trajectory.R
\name{remove_trajectory}
\alias{remove_trajectory}
\title{Remove the contribution of a subject from the compartment counts in the
\code{config_mat} and \code{obs_mat} objects.}
\usage{
remove_trajectory(epimodel, subject, save_path)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects.}

\item{subject}{ID for subject to be removed}

\item{save_path}{TRUE/FALSE for whether to save the current path in epimodel}
}
\value{
\code{config_mat} and \code{obs_mat} objects within the .epimodel
  environment.
}
\description{
Remove the contribution of a subject from the compartment counts in the
\code{config_mat} and \code{obs_mat} objects.
}

