% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reorderMat}
\alias{reorderMat}
\title{Reorder the rows of a matrix}
\usage{
reorderMat(oldmtx, ord)
}
\arguments{
\item{ord}{row order}

\item{mtx}{matrix to be reordered}
}
\value{
matrix with rows permuted according to ord
}
\description{
Reorder the rows of a matrix
}

