% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resolveSubjContrib}
\alias{resolveSubjContrib}
\title{Add or remove the subject contribution from the compartment counts}
\usage{
resolveSubjContrib(pop_mat, row_inds, col_inds, insertion)
}
\arguments{
\item{pop_mat}{population bookkeeping matrix in epimodel list}

\item{row_inds}{vector of row indices, 1:epimodel$ind_final_config}

\item{col_inds}{column inds given by the subject path in epimodel$config_mat}

\item{insertion}{Logical indicating whether the resolution should be the
    insertion or the removal of a path}
}
\value{
updated compartment counts
}
\description{
Add or remove the subject contribution from the compartment counts
}

