% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_at_event_times.R
\name{sample_at_event_times}
\alias{sample_at_event_times}
\title{Sample the infection status at times of state transition of other subjects}
\usage{
sample_at_event_times(epimodel, subject)
}
\value{
updated configuration matrix of subject level trajectories
}
\description{
Sample the infection status at times of state transition of other subjects
}

