% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_at_event_times.R
\name{sample_at_event_times}
\alias{sample_at_event_times}
\title{Sample the infection status at times of state transition of other subjects}
\usage{
sample_at_event_times(path, epimodel)
}
\arguments{
\item{path}{vector corresponding to a subject path.}

\item{epimodel}{list of bookkeeping and model objects.}
}
\value{
updated path vector of subject level trajectories
}
\description{
Sample the infection status at times of state transition of other subjects
}
