% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_at_obs_times.R
\name{sample_at_obs_times}
\alias{sample_at_obs_times}
\title{Sample the infection status for the specified subject at observation times.}
\usage{
sample_at_obs_times(config_mat, epimodel, subject)
}
\value{
updated configuration matrix
}
\description{
Sample the infection status for the specified subject at observation times.
}

