% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_DT_skeleton.R
\name{sample_DT_skeleton}
\alias{sample_DT_skeleton}
\title{Sample the state at event times from the discete time skeleton.}
\usage{
sample_DT_skeleton(path, epimodel, init_ind, final_ind)
}
\arguments{
\item{path}{column from the config_mat corresponding to a subject path}

\item{epimodel}{list of bookkeeping and model objects.}

\item{init_ind, final_ind}{index in the configuration matrix of the left/right
endpoints of the observation interval}

\item{subject}{subject ID, i.e. column index in config_mat}
}
\value{
updated path vector
}
\description{
This function only gets called when there is at least one transition event in
an inter-observation interval
}

