% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_forward.R
\name{sample_forward}
\alias{sample_forward}
\title{Forward path sampling.}
\usage{
sample_forward(epimodel, subject, subj_ID, init_time, final_time, init_state,
  final_state, irm_key)
}
\arguments{
\item{init_time}{}

\item{final_time}{}

\item{init_state}{}

\item{final_state}{}

\item{irm_key}{}
}
\value{
updated configuration matrix with a valid path (no update if constant
  path)
}
\description{
Forward path sampling.
}

