% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_path.R
\name{sample_path}
\alias{sample_path}
\title{Wrapper for sampling the path within homogeneous intervals.}
\usage{
sample_path(epimodel, subject)
}
\arguments{
\item{subject}{character string for subject ID code}
}
\value{
updated configuration matrix within the epimodel environment.
}
\description{
Wrapper for sampling the path within homogeneous intervals.
}

