% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_path_in_interval.R
\name{sample_path_in_interval}
\alias{sample_path_in_interval}
\title{Draw a path within an interval from an endpoint conditioned homogeneous CTMC.}
\usage{
sample_path_in_interval(epimodel, subject, interval)
}
\arguments{
\item{interval}{index for the interval within which path is to be drawn}
}
\value{
updated configuration matrix with subject path (excluding at time 0
  and tmax) at the end
}
\description{
Draw a path within an interval from an endpoint conditioned homogeneous CTMC.
}

