% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_one_event.R
\name{sim_one_event}
\alias{sim_one_event}
\title{Simulate the next event using Gillespie's next reaction method.}
\usage{
sim_one_event(config_list, epimodel)
}
\arguments{
\item{config_list}{list with three named elements - config, rate_mat, and
dt_mat - where config should correspond to a row in
\code{epimodel$config}.}

\item{epimodel}{list of bookkeeping and model objects. If not supplied, the
list will be generated.}
}
\value{
updated config_list
}
\description{
Internal function for \code{simulate_epimodel}, which is its parent
environment. This function uses the rate_mat, t, and epimodel objects in its
parent environment, modifying them there.
}

