% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_one_event.R
\name{sim_one_event}
\alias{sim_one_event}
\title{Simulate the next event using Gillespie's next reaction method.}
\usage{
sim_one_event(config_list, epimodel, lump)
}
\arguments{
\item{config_list}{list with three named elements - config_mat, pop_mat,
rate_mat, and dt_mat.}
}
\value{
updated config_list
}
\description{
Internal function for \code{simulate_epimodel}, which is its parent
environment. This function uses the rate_mat, t, and epimodel objects in its
parent environment, modifying them there.
}
