% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_one_event.R
\name{sim_one_event}
\alias{sim_one_event}
\title{Simulate the next event using Gillespie's next reaction method.}
\usage{
sim_one_event(config_list, epimodel, lump)
}
\arguments{
\item{config_list}{list with three named elements - config, rate_mat, and
dt_mat - where config should correspond to a row in
\code{epimodel$config}.}

\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}

\item{lump}{option specifying whether to simulate the next event directly on
  the extended state space of individuals, or to simulate on the lumped state
  space of compartment counts, and then choose a subject uniformly at random
  for each event from the set of subjects in the risk pool for that event. A
  value of TRUE, corresponding to the lumped state space, is appropriate when
  there is no subject level heterogeneity and is significantly faster in even
  moderate sized populations.}
}
\value{
updated config_list
}
\description{
Internal function for \code{simulate_epimodel}, which is its parent
environment. This function uses the rate_mat, t, and epimodel objects in its
parent environment, modifying them there.
}

