% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subjectLikelihood}
\alias{subjectLikelihood}
\title{Get the irm keys for the compartment counts in a population level
bookkeeping matrix}
\usage{
subjectLikelihood(subject, pop_mat, subj_path, irm_array, initdist, keys,
  loglik)
}
\arguments{
\item{pop_mat}{population level bookkeeping matrix}

\item{subj_path}{integer vector giving the subject path}

\item{irm_array}{array of rate matrices}

\item{keys}{vector of irm keys}

\item{loglik}{boolean indicating whether to return the likelihood or
log-likelihood}
}
\value{
subject level likelihood or log-likelihood
}
\description{
Get the irm keys for the compartment counts in a population level
bookkeeping matrix
}

