% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_to_obs.R
\name{traj_to_obs}
\alias{traj_to_obs}
\title{Instatiates a data matrix based on a configuration matrix. The code in this}
\usage{
traj_to_obs(epimodel = NULL)
}
\value{
data matrix for compartment counts at observation times.
}
\description{
Instatiates a data matrix based on a configuration matrix. The code in this
}
