% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traj_to_obs.R
\name{traj_to_obs}
\alias{traj_to_obs}
\title{Instatiates a data matrix based on a configuration matrix. The code in this}
\usage{
traj_to_obs(epimodel = NULL)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}
}
\value{
data matrix for compartment
}
\description{
Instatiates a data matrix based on a configuration matrix. The code in this
}

