% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_proposal.R
\name{tune_proposal}
\alias{tune_proposal}
\title{Tune proposal distribution for randow walk M-H parameter proposals.}
\usage{
tune_proposal(epimodel, epimodel_envir)
}
\arguments{
\item{epimodel_envir}{logical indicating whether the tuning is run within the
epimodel environment}
}
\value{
updated covariance matrix and scale factor within the epimodel 
  environment, or a list with tuning results
}
\description{
The empirical covariance matrix is tuned either by taking weighted averages
or by re-tuning the scale multiple until the acceptance rate is within the
desired range.
}
