% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_matrices.R
\name{update_matrices}
\alias{update_matrices}
\title{Update tpms, tpm products, emission matrices, and FB matrices before drawing
a new subject path.}
\usage{
update_matrices(epimodel, subject)
}
\value{
updated matrices within the epimodel environment.
}
\description{
Update tpms, tpm products, emission matrices, and FB matrices before drawing
a new subject path.
}

