% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_obs_mat.R
\name{update_obs_mat}
\alias{update_obs_mat}
\title{Update the observation matrix and the indices of observation times.}
\usage{
update_obs_mat(epimodel)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects, or an analagous
environment. If not supplied, the list will be generated. The epimodel list
is coerced into an environment, .epimodel, for internal use.}
}
\value{
updated observation matrix and obs_time_inds vector in the epimodel
  environment
}
\description{
Update the observation matrix and the indices of observation times.
}

