% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_obs_mat.R
\name{update_obs_mat}
\alias{update_obs_mat}
\title{Update the observation bookkeeping matrix with a new trajectory.}
\usage{
update_obs_mat(epimodel, direction, trajec)
}
\arguments{
\item{epimodel}{list of bookkeeping and model objects. If not supplied, the
list will be generated.}

\item{direction}{update observation matrix to reflect the addition or removal
of a subject-level trajectory. Valid arguments are \code{add} or
\code{remove}.}

\item{trajec}{matrix of configurations for a single subject, where each row
  is a row from \code{epimodel$config_mat} for that subject.}
}
\value{
updated \code{obs_mat} matrix
}
\description{
Update the observation bookkeeping matrix with a new trajectory.
}

